Refactor HEIC to HEIF.
